Extract base class DateField
